Selector stuff

+++ Assets/Scripts/Common/Data/IntData.cs
@@ -6,8 +6,8 @@ using UnityEngine;
 public class IntData : ScriptableObject
 {
 	[SerializeField] int _value;
-	public float min;
-	public float max;
+	public int min;
+	public int max;
 
 	public int value { get => _value; set => _value = value; }
 }
